Extra Sapace in closed form solution code files, file structure was fixed
